Margaret edit 3

+++ app.py
@@ -33,12 +33,36 @@ def game():
 
         return Response(status=201)
 
-@app.route('/api/games/<int:gameID>', methods=["DELETE"])
+@app.route('/api/games/<int:gameID>', methods=["DELETE", "PUT"])
 def delete_game(gameID):
     db_connection = db.connect_to_database()
-    query = "DELETE FROM Games WHERE gameID = %s;"
-    db.execute_query(db_connection=db_connection, query=query, query_params=(gameID,))
-    return Response(status=204)
+    if request.method == 'DELETE':
+        query = "DELETE FROM Games WHERE gameID = %s;"
+        db.execute_query(db_connection=db_connection, query=query, query_params=(gameID,))
+        return Response(status=204)
+    elif request.method == 'PUT':
+        form_data = request.get_json()
+        print("update game")
+        query = "UPDATE Games SET name=%s, min_number_player=%s, max_number_player=%s WHERE gameID=%s;"
+        db.execute_query(db_connection=db_connection, query=query, query_params=(form_data['name'], form_data['min_number_player'], form_data['max_number_player'], form_data['gameID']))
+        return Response(status=200)
+
+
+#@app.route('/api/games/<int:gameID>', methods=["PUT", "POST", "GET"])
+#def edit_game(gameID):
+#    print("gameID", gameID)
+#    db_connection = db.connect_to_database()
+#    if request.method == 'POST':
+##        print("update game")
+#        query = "UPDATE Games SET name=%s, min_number_player=%s, max_number_player=%s WHERE gameID=%s;"
+#        cursor = db.execute_query(db_connection=db_connection, query=query, query_params=(form_data['name'], form_data['min_number_player'], form_data['max_number_player'], form_data['gameID']))
+#        return Response(status=200)
+#    elif request.method =='PUT':
+#        print("update game")
+#        query = "UPDATE Games SET name=%s, min_number_player=%s, max_number_player=%s WHERE gameID=%s;"
+#        cursor = db.execute_query(db_connection=db_connection, query=query, query_params=(form_data['name'], form_data['min_number_player'], form_data['max_number_player'], form_data['gameID']))
+#        return Response(status=200)
+
 
 """ Palyers Page API """
 @app.route('/api/players', methods=["POST", "GET"])
